[revert] moved collapsed state into ExpandableTagsQuery
