+++ src/exampleTypes/TagsQuery/index.js
@@ -1,15 +1,39 @@
 import React from 'react'
 import { contexturifyWithoutLoader } from '../../utils/hoc'
 import ExpandableTagsQuery from '../ExpandableTagsQuery'
+import OutsideClickHandler from 'react-outside-click-handler'
+import F from 'futil'
+import _ from 'lodash/fp'
+import ExpandableTagsInput, { Tags } from '../../greyVest/ExpandableTagsInput'
 
-let TagsQuery = ({ tree, node, actionWrapper, ...props }) => (
-  <div className="tags-query" style={{ marginBottom: 28 }}>
-    <ExpandableTagsQuery
-      {...{ tree, node, actionWrapper, ...props }}
-      Loader={({ children }) => <div>{children}</div>}
-      style={{ padding: '0 5px' }}
-    />
-  </div>
-)
+let TagsQuery = ({ tree, node, actionWrapper, ...props }) => {
+  let collapse = React.useState(true)
+  let isCollapsed = F.view(collapse) && !_.isEmpty(node.tags)
+  return (
+    <OutsideClickHandler
+      onOutsideClick={() => {
+        F.on(collapse)()
+      }}
+      useCapture={false}
+    >
+      <div
+        className="tags-query"
+        onClick={F.off(collapse)}
+        style={{ marginBottom: isCollapsed ? 28 : 10 }}
+      >
+        <ExpandableTagsQuery
+          {...{ tree, node, collapse, actionWrapper, ...props }}
+          autoFocus
+          onAddTag={F.off(collapse)}
+          Loader={({ children }) => <div>{children}</div>}
+          style={{ padding: '0 5px' }}
+          theme={{
+            TagsInput: isCollapsed ? Tags : ExpandableTagsInput,
+          }}
+        />
+      </div>
+    </OutsideClickHandler>
+  )
+}
 
 export default contexturifyWithoutLoader(TagsQuery)
